requirements.txt
accelerate==1.3.0
aiosqlite==0.21.0
bcrypt==4.3.0
black==24.8.0
datasets==3.2.0
evaluate==0.4.3
fastapi==0.115.0
flake8==7.1.1
openai==1.51.0
pre-commit==3.8.0
python-dotenv==1.0.1
uvicorn==0.31.0
polars==1.9.0
pandas==2.2.3
passlib==1.7.4
pyarrow==17.0.0
requests==2.32.3
praw==7.8.1
pyjwt==2.10.1
scikit-learn==1.5.2
scipy==1.14.1
sqlalchemy==2.0.37
torch==2.5.1
transformers==4.48.0
protobuf==6.30.2
sentencepiece==0.2.0
xlsxwriter==3.2.5
